requirements.txt
asgiref==3.8.1
Babel==2.15.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-ckeditor-5==0.2.12
django-haystack==3.3.0
django-js-asset==2.2.0
django-tagging==0.5.0
django-taggit==5.0.1
django-tagulous==1.3.3
django-watson==1.6.3
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
et-xmlfile==1.1.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
markdown2==2.4.6
httplib2==0.22.0
idna==3.7
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2